framework-examples/nemo-agno-personal-finance-agent/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xpander_nemo_agent"
version = "0.0.1"
description = "xpander.ai Nvidia NeMo + Agno"
authors = [{name = "xpander.ai", email = "info@xpander.ai"}]
requires-python = ">=3.11"
dependencies = [
    "python-dotenv",
    "agno>=1.8.2",
    "xpander-sdk[agno]", 
    "openai",
    "anthropic",
    "mcp",
    "nvidia-nat",
    "loguru",
    "build",
    "litellm",
    "google-search-results"
]

[tool.setuptools]
py-modules = ["xpander_nemo_agent"]

[project.entry-points."nat.plugins"]
xpander_nemo_agent = "xpander_nemo_agent"